refactor: Centralize UPS service code definitions, aliases, and translation logic into a dedicated `ups_service_codes` module.
